move streamToString
