+++ e2e/baseFixtures.js
@@ -171,19 +171,5 @@ exports.test = base.test.extend({
     }
 });
 
-/**
- * Takes a readable stream and returns a string.
- * @param {ReadableStream} readable - the readable stream
- * @return {Promise<String>} the stringified stream
- */
-exports.streamToString = async function (readable) {
-    let result = '';
-    for await (const chunk of readable) {
-        result += chunk;
-    }
-
-    return result;
-};
-
 exports.expect = expect;
 exports.waitForAnimations = waitForAnimations;
